Add files via upload
This is the attempt to replicate the heart-disease example in the book by James et al. on tree model. For fun, I compared the decision tree (that I had tuned with a large array of cp values) with a simple logistic regression model. Result: The logistic regression was far better in the training data as well as in the test set. The confusion matrices (test set) of both models showed that the decision tree had only a 50% chance to correctly predict a person with a heart disease....